27.8.Project end

+++ src/main/java/org/sf247/compare/CompareStudentAvgExamScore.java
@@ -10,6 +10,6 @@ import org.sf247.modelclass.Student;
 public class CompareStudentAvgExamScore implements CompareStudents{
     @Override
     public int compare(Student o1, Student o2) {
-        return Float.compare(o2.getAvgExamScore(), o1.getAvgExamScore());
+        return Double.compare(o2.getAvgExamScore(), o1.getAvgExamScore());
     }
 }
